Convert nightfall to typescript
